requirements.txt
arrow==1.2.3
langchain==0.0.287
nltk==3.8.1
openai==0.28.0
pandas==1.5.3
pdf2image==1.16.3
pdfminer==20191125
pdfminer.six==20221105
pdfplumber==0.10.2
pinecone-client==2.2.2
pyarrow==11.0.0
pyparsing==3.0.9
pypdfium2==4.20.0
python-dateutil==2.8.2
python-json-logger==2.0.6
questionary==1.10.0
six==1.16.0
tiktoken==0.5.1
unstructured==0.10.14
unstructured-inference==0.5.27

